backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
openai==1.3.7
requests==2.31.0
python-dateutil==2.8.2
